requirements.txt
streamlit
openai
chromadb
sentence-transformers
PyPDF2
python-docx
tiktoken
anthropic
pysqlite3-binary; sys_platform == 'linux'
